==> 03_FDOC.ipynb <==
Florida Department of Corrections
---------------------------------

The Florida Department of Corrections has made a number of tables available about inmates who have passed through their prisons. [The main publication page for the Offender Based Information System (OBIS)](http://www.dc.state.fl.us/pub/obis_request.html) lists what each table contains. The data are said to be underneath their [interactive web applications](http://www.dc.state.fl.us/InmateInfo/InmateInfoMenu.asp). Have a look at these descriptions and at the web search portals and comment on what you see.

> - **Inmate_release_root**: Basic inmate information on inmates released since 10-1-1997
- **Inmate_release_aliases**: Aliases for these released inmates
- **Inmate_release_offenses_CPS**: Current prison offenses for these released inmates
- **Inmate_release_offenses_prpr**: Prior prison offenses for these released inmates 
- **Inmate_release_residence**: Release plan address for these released inmates
- **Inmate_release_detainers**: Detainers for these released inmates
- **Inmate_release_incarhist**: Incarceration history for these released inmates
- **Inmate_release_scarsmarks**: Tattoos for these released inmates
- **Inmate_active_root**:  Basic inmate information on active inmates
- **Inmate_active_aliases**: Aliases for these active inmates
- **Inmate_active_offenses_CPS**: Current prison offenses for these active inmates
- **Inmate_active_offenses_prpr**: Prior prison offenses for these active inmates 
- **Inmate_active_detainers**: Detainers for these active inmates
- **Inmate_active_incarhist**: Incarceration history for these active inmates
- **Inmate_active_scarsmarks**: Tattoos for these active inmates
- **Offender_root**: Basic offender information on current community supervision offenders
- **Offender_aliases**: Aliases for offenders
- **Offender_offenses_CCS**: Current community supervision offenses for offenders
- **Offender_residence**: Most recent address (excluding confidential addresses) for offenders

You can also do some simple sleuthing using basic UNIX tools. First, I have made the data files available individually as well as bundled for easier downloading. Here is [the folder with the separate files](http://compute-cuj.org/FDOC/) and [FDOC.tar.gz](http://compute-cuj.org/FDOC/FDOC.tar.gz) is the tar'd (yes, a tape archive) and compressed folder. 

To have a look at these files, we can read in just a few observations and avoid long load times. 

library(dplyr)

tmp = read.csv("http://compute-cuj.org/FDOC/INMATE_ACTIVE_ROOT.csv",as.is=TRUE,nrows=20)
tmp

The FDOC commissions studies of their system based on these data. Some studies involve data we don't have access to, but others seem to be quite creative with the data that is here. Take [this report on recidivism](http://www.dc.state.fl.us/pub/recidivismfsu/index.html), for example. 

Let's read in the main "root" file for current inmates.

root = read.csv("http://compute-cuj.org/FDOC/INMATE_ACTIVE_ROOT.csv",as.is=TRUE)

table(root$FACILITY)

table(root$Race,root$Sex)

And have a look at a few more of the files...

tmp = read.csv("http://compute-cuj.org/FDOC/INMATE_ACTIVE_OFFENSES_prpr.csv",as.is=TRUE,nrows=20)
head(tmp)

tmp = read.csv("http://compute-cuj.org/FDOC/INMATE_ACTIVE_INCARHIST.csv",as.is=TRUE,nrows=20)
tmp

tmp = read.csv("http://compute-cuj.org/FDOC/INMATE_ACTIVE_SCARSMARKS.csv",as.is=TRUE,nrows=20)
tmp
