==> 07-Genderizer.ipynb <==
Brief jump to Python
--------------------

The same notebook can be used to run Python code. Granted the syntax is very different, the overall style of coding is the same. First, we install a package (like an R package) that attempts to asign gender to different first names. Yes, it's dicey but this program at least provides you an indicator of its certainty about a name.

First, install the package.

%%bash
pip install gender-guesser

The first line below is like R's library() command. This takes the data and objects we just created and loads them into our Python session for our use. We then create a Detector() and use it to assign genders to names...

from gender_guesser.detector import Detector
d = Detector()

print(d.get_gender(u"Rick"))
print(d.get_gender(u"Sally"))
print(d.get_gender(u"Pauley")) # should be androgynous

Not too bad! We will then read in our CSV file, take the first name, pass it through the detector and add the column to our CSV. Here we go! Notice that we are using a loop and that the syntax is not entirely dissimilar from R's.

from csv import reader, writer

infile = reader(open("newcomms.csv"))
outfile = writer(open("newcomms2.csv","w"))

row = next(infile)
row.insert(3,"gender")

outfile.writerow(row)

for row in infile:
    
    gender = d.get_gender(row[2])
    row.insert(3,gender)
    if len(row)!=22:    
        print(len(row))
        print(row)

    outfile.writerow(row)
