==> 10_Spies.ipynb <==
Analyzing spatial data
-----------------------

TosY we are going to consider an article that appeared in Buzzfeed by Peter Aldhous and Charles Seife. It is about surveillance flights piloted by the FBI and DHS. The article is called [Spies in the Skies](https://www.buzzfeed.com/peteraldhous/spies-in-the-skies). Please read the article, thinking about the ways data informed their narrative.

1. Where did the data come from? If you were given this assignment, could you assemble this data set? How?
2. What are the basic data they are working with and what kinds of analysis did they do?
3. The data and analysis are available in CSV format and R code. Where? 

Try to recreate a bit of their analysis using their documentation. What other questions would you want to ask?

**Preface: An unfamiliar construction**

The authors make use of a construction for orchestrating computations in R that we haven't covered yet. It's called a "pipe". It is a common construction in many languages, and you might recognize it from the UNIX shell. There is a nice [description of pipes in R here](https://www.r-statistics.com/2014/08/simpler-r-coding-with-pipes-the-present-and-future-of-the-magrittr-package/).

The symbol "%>%" is known as a pipe. It means taking the data that is output by one command and feeding it as input to the following command. So, if we want to filter() the crime data below to just "MANHATTAN" incidents and then sample_n() 25 of them, we might do the following.

library(dplyr)
library(ggplot2)
library(magrittr)  # ceci n'est pas un pipe

crime = read.csv("https://data.cityofnewyork.us/api/views/5uac-w243/rows.csv",as.is=TRUE)

crime %>%
  filter(BORO_NM=="MANHATTAN") %>%
  sample_n(25) 

So, to read this, you start with our crime data frame. It is taken as an input to the filter() command that looks just for "MANHATTAN". The output of this is then fed to the sample command. Notice that the first arugment to filter() and sample_n() and all of the "dplyr" functions is a data set. We drop reference to those data sets when using pipes because the data are taken from the previous command  in the "pipeline". 

The syntax is really catching on with R users. What do you think? 

Here we sample 10,000 points and then feed it to plot. When we don't have a "dplyr" function, we can use the "data=." argument to indicate that the data are to come from the output of the operation upstream. Here's a plot.

crime %>%
  filter(BORO_NM=="MANHATTAN") %>%
  sample_n(10000) %>%
  plot(Latitude~Longitude,data=.,pch=".")

Also, they a different assignment operator. We have been using the equality sign to create a variable...

x = 5

Old school R (back when it was S) used "<-" instead. The motivation was twofold. First, on an old AT&T Terminal (APL keyboard as well), there was a single key that represented an arrow sign. So it was as easy to type as "=". Second, if there was a convenient second symbol that could represent assignment, the authors of the language thought it would be easier for a novice to differentiate that "x <- 5" means assign the label "x" to the value 5; and that sample_n(size=5) means pass the value of 5 under the name "size" in the function "sample_n". 

The way I've taught you, "=" could mean create a new variable or pass a value into a function. It was felt that that could be confusing for beginners and "<-" was born. What do you think? 

So, you will often see people writing R code with "<-" as the assignment operator. Except in a couple of small cases, the two approaches, "<-" and "=", are equivalent when making assignments.

x <- 5

Aside from this, you should be able to read Peter's code. Read the article and try out their work!

Having found patterns of navigation, Peter decided to train a learning algorithm to find other planes that exhibited these patterns. Have a look at his next article, [BuzzFeed News Trained A Computer To Search For Hidden Spy Planes. This Is What We Found.](https://www.buzzfeed.com/peteraldhous/hidden-spy-planes?utm_term=.rg5odvnlp#.bfqYmZl12) We will spend the next few lectures talking about models like the ones he used in this article.

His data and code are available on github for this second article as well. Have a look and see what he did.
